qcacld-2.0: Fixing out-of-bound memory read while loading module
hdd_update_mac_config reads the contents of wlan_mac.bin by calling
the function request_firmware. It then passes the resulting buffer
to get_next_line which parses the buffer as a string. The problem
is that get_next_line expects the string to be NULL terminated or
have a END marker but the contents of the buffer is neither NULL
terminated nor having END marker. Thus when get_next_line gets to
the end of the buffer, it continues to read until it hits a NULL in
memory or until it hits the memory exception.
Same issue can exist in hdd_parse_config_ini function for qcom_cfg.ini
file hence fixing there as well.

Change-Id: I2c29db7a97399f807e6c2a428500d99000dc7a52
CRs-Fixed: 2115899