qcacld-2.0: Add sanity checks for FILS connection
propagation from qcacld-3.0 to qcacld-2.0

Add the sanity checks to the authentication types for a
FILS connection to ensure if it is a FILS type connection
and if it is, then add a check for the currently supported
FILS authentication types. This should also take care of
getting rid of dangling pointers for fils_con_info which
are not set to NULL after a memory release operation.

Change-Id: I688086fdff1536bf7e314832e34012878e3a3304
CRs-Fixed: 2229287